retry `ALTAIR_FORK_EPOCH` query in local testnet (+11 squashed commits) Squashed commits: [0db44990] formatting [736f9845] align variable name in template [58379772] fix LC data pruning for empty epoch boundary [42792616] fix LC data pruning for initial period [cdd87dbf] formatting [3a6f792c] avoid mixing code move with other changes [34a761b3] pass `bid` instead of full block to `cacheLightClientData` [8e5e8d30] remove unnecessary `withStateVars` wrapper [9c63cd54] `earliestSlot` -> `lowSlot` where unrelated to LC data [1ee30221] rm mentions of local variable in comment [2ebfdca0] preparations for persisting light client data
When switching the light client data in-memory cache to a database,
additional failure points are introduced due to possible inconsistencies
that are introduced into the DB out-of-band. These additional failure
points are handled as preparation to the move to persistence.
Also includes fixes related to theoretic edge cases in initial period,
and some minor cleanups, e.g., `asSigned` vs `isomorphicCast`.